split out utils for rest
