+++ services/web/server/src/simcore_service_webserver/projects/_crud_handlers_utils.py
@@ -1,0 +1,48 @@
+from aiohttp import web
+from models_library.api_schemas_webserver.projects import ProjectListItem
+from models_library.rest_pagination import Page
+from models_library.rest_pagination_utils import paginate_data
+from servicelib.mimetype_constants import MIMETYPE_APPLICATION_JSON
+from servicelib.rest_constants import RESPONSE_MODEL_POLICY
+
+from . import _permalink_api
+from ._crud_api_read import _paralell_update
+from .models import ProjectDict
+
+
+async def aggregate_data_to_projects_from_request(
+    request: web.Request,
+    projects: list[ProjectDict],
+) -> list[ProjectDict]:
+
+    update_permalink_per_project = [
+        # permalink
+        _permalink_api.aggregate_permalink_in_project(request, project=prj)
+        for prj in projects
+    ]
+
+    updated_projects: list[ProjectDict] = await _paralell_update(
+        *update_permalink_per_project,
+    )
+    return updated_projects
+
+
+def create_page_response(projects, request_url, total, limit, offset) -> web.Response:
+    page = Page[ProjectListItem].model_validate(
+        paginate_data(
+            chunk=[
+                ProjectListItem.from_domain_model(prj).model_dump(
+                    by_alias=True, exclude_unset=True
+                )
+                for prj in projects
+            ],
+            request_url=request_url,
+            total=total,
+            limit=limit,
+            offset=offset,
+        )
+    )
+    return web.Response(
+        text=page.model_dump_json(**RESPONSE_MODEL_POLICY),
+        content_type=MIMETYPE_APPLICATION_JSON,
+    )
